chore: added new AppError, refactored code and added validation to most APIs

+++ src/application/use-cases/task/DeleteTask.usecase.ts
@@ -1,10 +1,14 @@
-import { Task } from "../../../domain/entities/Task.entity";
 import { TaskRepository } from "../../../domain/repositories/TaskRepository.repo";
+import { AppError } from "../../../shared/utils/AppError";
 
 export class DeleteTaskUseCase {
-    constructor(private taskRepository: TaskRepository) {}
+    constructor(private taskRepository: TaskRepository) { }
 
-    async execute(taskId: number): Promise<Task> {
+    async execute(taskId: number) {
+        const taskExists = await this.taskRepository.getById(taskId);
+        if (!taskExists) {
+            throw new AppError("Task does not exists", 404)
+        }
         return this.taskRepository.delete(taskId);
     }
 }
