+++ src/application/use-cases/task_assignment/DeleteTaskAssignment.usecase.ts
@@ -1,3 +1,4 @@
+import { AppError } from '../../../shared/utils/AppError';
 import { TaskAssignmentRepository } from './../../../domain/repositories/TaskAssignmentRepository.repo';
 export class DeleteTaskAssignmentUseCase {
     private taskAssignmentRepo: TaskAssignmentRepository
@@ -5,10 +6,16 @@ export class DeleteTaskAssignmentUseCase {
         this.taskAssignmentRepo = TaskAssignmentRepository
     }
 
-    execute(ta_id: number) {
+    async execute(ta_id: number) {
         if (!ta_id) {
-            throw new Error("Missing required filed task_assignment_id!")
+            throw new AppError("Missing required filed task_assignment_id!", 400)
         }
+
+        const tAssnExists = await this.taskAssignmentRepo.getById(ta_id)
+        if (!tAssnExists) {
+            throw new AppError(`Task Assignment not found!`, 404)
+        }
+
         return this.taskAssignmentRepo.delete(ta_id)
     }
 }
